03_Vector_Databases/FAISS/requirements.txt
# Use faiss-gpu if you have a compatible NVIDIA GPU and CUDA installed
faiss-cpu
numpy # FAISS often requires numpy arrays
sentence-transformers
openai
python-dotenv
pandas 